Refactor common code across IMSC reader and writer (#43)
